feature: show trip logs on map
