+++ frontend/src/types/Touring.ts
@@ -1,0 +1,16 @@
+export type TripTypes = {
+  lat: number;
+  lng: number;
+};
+
+export type PhotoTypes = {
+  id: string;
+  lat: number;
+  lng: number;
+};
+
+export type TouringLogTypes = {
+  trip: TripTypes[];
+  photo: PhotoTypes[];
+  center: TripTypes;
+};
